quote order created

+++ INTEX/Models/Discount.cs
@@ -23,7 +23,7 @@ namespace INTEX.Models
 
         [DisplayName("Discount Percentage")]
         [Required(ErrorMessage = "Please enter the discount percentage")]
-        public double percentDiscount { get; set; }
+        public decimal percentDiscount { get; set; }
 
     }
 }
